requirements.txt
pandas
plotly.express
streamlit
plotly.graph_objects